% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{encodeViralSeq}
\alias{encodeViralSeq}
\title{Viral Protein Sequence Encoding with GloVe Embedding Vectors}
\usage{
encodeViralSeq(trainingSet, embeddings_index)
}
\arguments{
\item{trainingSet}{a data.frame containing training information}

\item{embeddings_index}{embedding outputted from
\code{\link[DeProViR]{gloveImport}}}
}
\value{
A list containing Embedding matrix and tokenization
}
\description{
This function first first encodes amino acids as a sequence
of unique 20 integers though tokenizer. The padding token was added to the
front of shorter sequences to ensure a fixed-length vector of
defined size L (i.e., here is 1000). Embedding matrix is then constructed
to transform amino acid tokens to pre-training embedding weights, in which
rows represent the amino acid tokens created earlier, and columns
correspond to 100-dimension weight vectors derived from GloVe
word-vector-generationvector map.
}
\examples{
# Download and load the index
embeddings_index <- gloveImport()
#load training set
dt <- loadTrainingSet()
#encoding
encoded_seq <- encodeViralSeq(dt, embeddings_index)
}
