% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gloveImport}
\alias{gloveImport}
\title{Cache and Load Pre-Trained Word Vectors}
\usage{
gloveImport(url_path = "https://nlp.stanford.edu/data")
}
\arguments{
\item{url_path}{URL path to GloVe embedding. Defaults to
"https://nlp.stanford.edu/data"}
}
\value{
glove embedding
}
\description{
This function cache and loads pre-trained GloVe
vectors (100d).
}
\examples{
options(timeout=240)
embeddings_index <-
gloveImport(url_path = "https://nlp.stanford.edu/data")
}
