% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Load_PreTrainedModel}
\alias{Load_PreTrainedModel}
\title{Load Pre-Trained Model Wights}
\usage{
Load_PreTrainedModel(
  input_dim = 20,
  output_dim = 100,
  filters_layer1CNN = 32,
  kernel_size_layer1CNN = 16,
  filters_layer2CNN = 64,
  kernel_size_layer2CNN = 7,
  pool_size = 30,
  layer_lstm = 64,
  units = 8,
  metrics = "AUC",
  filepath = system.file("extdata", "Pre_trainedModel", package = "DeProViR")
)
}
\arguments{
\item{input_dim}{Integer. Size of the vocabulary, i.e. amino acid tokens.
Defults to 20. See \code{keras}.}

\item{output_dim}{Integer. Dimension of the dense embedding, i.e., GloVe.
Defaults to 100. See \code{keras}.}

\item{filters_layer1CNN}{Integer, the dimensionality of the output space
(i.e. the number of output filters in the first convolution).
Defaults to 32. See \code{keras}}

\item{kernel_size_layer1CNN}{An integer or tuple/list of 2 integers,
specifying the height and width of the convolution window in the first
layer. Can be a single integer to
specify the same value for all spatial dimensions.
Defaults to 16. See \code{keras}}

\item{filters_layer2CNN}{Integer, the dimensionality of the output space
(i.e. the number of output filters in the second convolution).
Defaults to 64. See \code{keras}}

\item{kernel_size_layer2CNN}{An integer or tuple/list of 2 integers,
specifying the height and width of the convolution window in the
second layer. Can be a single integer to
specify the same value for all spatial dimensions.
Defaults to 7. See \code{keras}}

\item{pool_size}{Down samples the input representation by taking the
maximum value over a spatial window of size pool_size.
Defaults to 30.See \code{keras}}

\item{layer_lstm}{Number of units in the Bi-LSTM layer. Defaults to 64.
See \code{keras}}

\item{units}{Number of units in the MLP layer. Defaults to 8.
See \code{keras}}

\item{metrics}{Vector of metric names to be evaluated by the model during
training and testing. Defaults to "AUC". See \code{keras}}

\item{filepath}{A character string indicating the path contained
pre-trained model weights, i.e., inst/extdata/Pre-trainedModel}
}
\value{
Pre-trained model.
}
\description{
This function loads the pre-trained model weights constructed
previously using \code{\link[DeProViR]{ModelTraining}}
}
\details{
Load_PreTrainedModel
}
\examples{
Loading_trainedModel <- Load_PreTrainedModel()
}
