% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ModelPerformance_evalPlots}
\alias{ModelPerformance_evalPlots}
\title{Model Performance Evalution}
\usage{
ModelPerformance_evalPlots(pred_label, y_label, tpath = tempdir())
}
\arguments{
\item{pred_label}{predicted labels}

\item{y_label}{Ground truth labels}

\item{tpath}{A character string indicating the path to the project
directory. If the directory is
missing, PDF file will be stored in the Temp directory.}
}
\value{
Pdf file containing perfromanc plots
}
\description{
This function plots model performance
}
\details{
ModelPerformance_evalPlots
}
\examples{
pred_label = seq(0,1, length.out = 100)
truth_label = rep(c(0,1), each = 50)
perf <- ModelPerformance_evalPlots(pred_label, truth_label,
tpath = tempdir())
}
