% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ModelTraining}
\alias{ModelTraining}
\title{Predictive Model Training using k-fold Validation Strategy}
\usage{
ModelTraining(
  url_path = "https://nlp.stanford.edu/data",
  training_dir = system.file("extdata", "training_Set", package = "DeProViR"),
  input_dim = 20,
  output_dim = 100,
  filters_layer1CNN = 32,
  kernel_size_layer1CNN = 16,
  filters_layer2CNN = 64,
  kernel_size_layer2CNN = 7,
  pool_size = 30,
  layer_lstm = 64,
  units = 8,
  metrics = "AUC",
  cv_fold = 10,
  epochs = 100,
  batch_size = 128,
  plots = TRUE,
  tpath = tempdir(),
  save_model_weights = TRUE,
  filepath = tempdir()
)
}
\arguments{
\item{url_path}{URL path to GloVe embedding. Defaults to
"https://nlp.stanford.edu/data/glove.6B.zip".}

\item{training_dir}{dir containing viral-host training set.
See \code{\link[DeProViR]{load_TrainingSet}}}

\item{input_dim}{Integer. Size of the vocabulary, i.e. amino acid
tokens. Defults to 20. See \code{keras}.}

\item{output_dim}{Integer. Dimension of the dense embedding,
i.e., GloVe. Defaults to 100. See \code{keras}.}

\item{filters_layer1CNN}{Integer, the dimensionality of the output space
(i.e. the number of output filters in the first convolution).
Defaults to 32. See \code{keras}}

\item{kernel_size_layer1CNN}{An integer or tuple/list of 2 integers,
specifying the height and width of the convolution window in the first
layer. Can be a single integer to specify the same value for all
spatial dimensions.Defaults to 16. See \code{keras}}

\item{filters_layer2CNN}{Integer, the dimensionality of the output space
(i.e. the number of output filters in the second convolution).
Defaults to 64. See \code{keras}}

\item{kernel_size_layer2CNN}{An integer or tuple/list of 2 integers,
specifying the
height and width of the convolution window in the second layer. Can be a
single integer to specify the same value for all spatial dimensions.
Defaults to 7. See \code{keras}}

\item{pool_size}{Down samples the input representation by taking the
maximum value over a spatial window of size pool_size.
Defaults to 30.See \code{keras}}

\item{layer_lstm}{Number of units in the Bi-LSTM layer. Defaults to 64.
See \code{keras}}

\item{units}{Number of units in the MLP layer. Defaults to 8.
See \code{keras}}

\item{metrics}{Vector of metric names to be evaluated by the model
during training and testing. Defaults to "AUC". See \code{keras}}

\item{cv_fold}{Number of partitions for cross-validation. Defaults to 10.}

\item{epochs}{Number of epochs to train the model. Defaults to 100.
See \code{keras}}

\item{batch_size}{Number of samples per gradient update.Defults to 128.
See \code{keras}}

\item{plots}{PDF file containing perfromance measures. Defaults to TRUE.
See \code{\link[DeProViR]{ModelPerformance_evalPlots}}}

\item{tpath}{A character string indicating the path to the project
directory. If the directory is missing, PDF file containing perfromance
measures will be stored in the Temp directory.
See \code{\link[DeProViR]{ModelPerformance_evalPlots}}}

\item{save_model_weights}{If TRUE, save the trained weights.
Defaults to TRUE.}

\item{filepath}{A character string indicating the path to save
the model weights. Default to tempdir().}
}
\value{
Trained model and perfromance measures.
}
\description{
This function first transforms protein sequences to
amino acid tokens wherein tokens are indexed by positive integers,
then represents each amino acid token by pre-trained co-occurrence
embedding vectors learned by GloVe,
followed by applying an embedding layer. Then it employs Siamese-like
neural network articheture on densly-connected neural net to
predict interactions between host and viral proteins.
}
\details{
ModelTraining
}
