% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{predInteractions}
\alias{predInteractions}
\title{Predict Unknown Interactions}
\usage{
predInteractions(glove_dir, Testingset, trainedModel)
}
\arguments{
\item{glove_dir}{dir containing pre-trained word vectors (100d)}

\item{Testingset}{A data.frame containing unknown interactions. For demo,
we can use the file in extdata/test_Set.}

\item{trainedModel}{Pre-trained model stored in extdata/Pre_trainedModel
or the training model "$merge_model" achieved by
\code{\link[DeProViR]{ModelTraining}}.}
}
\value{
Probability scores for unknown interactions
}
\description{
This function initially constructs an embedding matrix from
the viral or host protein sequences and then predicts scores for unknown
interactions. Interactions with scores greater than 0.5 are more likely
to indicate interaction.
}
\details{
predInteractions
}
\examples{
#download glove index
zip_url <- "https://nlp.stanford.edu/data/glove.6B.zip"
#Destination folder to save the downloaded file
dest_folder = tempdir()
#Download the zip file
download.file(zip_url, destfile = file.path(dest_folder, "glove.6B.zip"),
mode = "wb")
# Unzip only the glove.6B.100d.txt file
unzip(zipfile = file.path(dest_folder, "glove.6B.zip"),
files = "glove.6B.100d.txt", exdir = dest_folder)
if (keras::is_keras_available() & reticulate::py_available()){
# load pre-trained model
trainedModel <- Load_PreTrainedModel()
# load test set (i.e., unknown interactions)
testing_set <- data.table::fread(
system.file("extdata/test_set/test_set_unknownInteraction.csv",
package = "DeProViR"))
# now predict interactions
predInteractions <-
 predInteractions(glove_dir = dest_folder, testing_set, trainedModel)
}

}
