% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gloveEmb_import}
\alias{gloveEmb_import}
\title{Cache and Load Pre-Trained Word Vectors}
\usage{
gloveEmb_import(url_path = "https://nlp.stanford.edu/data")
}
\arguments{
\item{url_path}{URL path to GloVe embedding. Defaults to
"https://nlp.stanford.edu/data"}
}
\value{
glove embedding
}
\description{
This function cache and loads pre-trained GloVe
vectors (100d).
}
\details{
gloveEmb_import
}
\examples{
options(timeout=240)
embeddings_index <-
gloveEmb_import(url_path = "https://nlp.stanford.edu/data")
}
